+++ crates/swift-bridge-ir/src/parse/parse_extern_mod.rs
@@ -805,158 +805,6 @@ mod tests {
         assert_eq!(parse_errors(tokens).len(), 0,);
     }
 
-    /// Verify that we can parse the `already_declared` attribute.
-    #[test]
-    fn parse_already_declared_attribute() {
-        let tokens = quote! {
-            mod foo {
-                extern "Rust" {
-                    #[swift_bridge(already_declared)]
-                    type AnotherType;
-                }
-            }
-        };
-
-        let module = parse_ok(tokens);
-
-        assert!(
-            module
-                .types
-                .get("AnotherType")
-                .unwrap()
-                .unwrap_opaque()
-                .attributes
-                .already_declared
-        );
-    }
-
-    //Verify that we can parse the `hashable` attribute.
-    #[test]
-    fn parse_hashable_attribute() {
-        let tokens = quote! {
-            mod foo {
-                extern "Rust" {
-                    #[swift_bridge(Hashable)]
-                    type SomeType;
-                }
-            }
-        };
-
-        let module = parse_ok(tokens);
-
-        assert_eq!(
-            module
-                .types
-                .get("SomeType")
-                .unwrap()
-                .unwrap_opaque()
-                .attributes
-                .hashable,
-            true
-        );
-    }
-
-    /// Verify that we can parse the `equatable` attribute.
-    #[test]
-    fn parse_equatable_attribute() {
-        let tokens = quote! {
-            mod foo {
-                extern "Rust" {
-                    #[swift_bridge(Equatable)]
-                    type SomeType;
-                }
-            }
-        };
-
-        let module = parse_ok(tokens);
-
-        assert_eq!(
-            module
-                .types
-                .get("SomeType")
-                .unwrap()
-                .unwrap_opaque()
-                .attributes
-                .equatable,
-            true
-        );
-    }
-
-    /// Verify that we can parse the `copy` attribute.
-    #[test]
-    fn parse_copy_attribute() {
-        let tokens = quote! {
-            mod foo {
-                extern "Rust" {
-                    #[swift_bridge(Copy(4))]
-                    type SomeType;
-                }
-            }
-        };
-
-        let module = parse_ok(tokens);
-
-        assert_eq!(
-            module
-                .types
-                .get("SomeType")
-                .unwrap()
-                .unwrap_opaque()
-                .attributes
-                .copy
-                .unwrap()
-                .size_bytes,
-            4
-        );
-    }
-
-    /// Verify that we can parse multiple atributes from an opaque type.
-    #[test]
-    fn parse_multiple_attributes() {
-        let tokens = quote! {
-            mod foo {
-                extern "Rust" {
-                    #[swift_bridge(already_declared, Copy(4))]
-                    type SomeType;
-                }
-            }
-        };
-
-        let module = parse_ok(tokens);
-
-        let ty = module.types.get("SomeType").unwrap().unwrap_opaque();
-        assert!(ty.attributes.copy.is_some());
-        assert!(ty.attributes.already_declared)
-    }
-
-    /// Verify that we can parse a doc comment from an extern "Rust" opaque type.
-    #[test]
-    fn parse_opaque_rust_type_doc_comment() {
-        let tokens = quote! {
-            mod foo {
-                extern "Rust" {
-                    /// Some comment
-                    type AnotherType;
-                }
-            }
-        };
-
-        let module = parse_ok(tokens);
-
-        assert_eq!(
-            module
-                .types
-                .get("AnotherType")
-                .unwrap()
-                .unwrap_opaque()
-                .attributes
-                .doc_comment
-                .as_ref()
-                .unwrap(),
-            " Some comment"
-        );
-    }
-
     /// Verify that we push errors for unknown arguments in a function
     #[test]
     fn error_args_into_arg_not_found_in_function() {
